+++ src/charts/fixtures/bar-chart-category-mix-many.js
@@ -1,0 +1,14 @@
+export default [
+  { y: 2600, name: 'Jeans' },
+  { y: 1900, name: 'Tops' },
+  { y: 1625, name: 'Shirts' },
+  { y: 1475, name: 'Cardigans' },
+  { y: 1000, name: 'Pull-overs' },
+  { y: 800, name: 'Jackets and tops' },
+  { y: 400, name: 'Chino pants' },
+  { y: 200, name: 'Ridiculous White Board Fancy Irregular Shorts' },
+  { y: 100, name: 'Hats' },
+  { y: 90, name: 'Panamas' },
+  { y: 80, name: 'Accessories' },
+  { y: 50, name: 'Jewelry' },
+];