+++ src/charts/fixtures/column-chart.js
@@ -1,9 +1,7 @@
-export const categories = [
-  ['Immediate', ''],
-  ['Jan 15, 2019—', 'Jan 28, 2019'],
-  ['Feb 15, 2019—', 'Mar 28, 2019'],
-  ['Apr 15, 2019—', 'May 28, 2019'],
-  ['Jun 15, 2019—', 'Jul 28, 2019'],
+export default [
+  { y: 231, name: ['Immediate', ''] },
+  { y: 91, name: ['Jan 15, 2019—', 'Jan 28, 2019'] },
+  { y: 321, name: ['Feb 15, 2019—', 'Mar 28, 2019'] },
+  { y: 81, name: ['Apr 15, 2019—', 'May 28, 2019'] },
+  { y: 263, name: ['Jun 15, 2019—', 'Jul 28, 2019'] },
 ];
-
-export const data = [231, 91, 321, 81, 263];