+++ src/globals/styles.js
@@ -1,11 +1,36 @@
 import { injectGlobal } from 'styled-components';
 import 'what-input';
-import { normalize } from '../theme';
+import { normalize, theme } from '../theme';
+
+const fontFaces = `
+@font-face {
+  font-family: 'Avenir Next';
+  src: url('/fonts/Avenir/AvenirNext-Light.ttf');
+  font-weight: 100;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: 'Avenir Next';
+  src: url('/fonts/Avenir/AvenirNext-Regular.ttf');
+  font-weight: 400;
+  font-style: normal;
+}
+
+
+@font-face {
+  font-family: 'Avenir Next';
+  src: url('/fonts/Avenir/AvenirNext-Medium.ttf');
+  font-weight: 500;
+  font-style: normal;
+}
+`;
 
 const injectGlobalStyles = () => injectGlobal`
+  ${fontFaces}
   ${normalize};
   html, body {
-    font-family: "Avenir Next", Arial, "Helvetica Neue", Helvetica, sans-serif;
+    font-family: ${theme.fontFamily};
     font-weight: normal;
   }
   p > img:not([class]) {